Change string.Repeat to use SpanAction string.Create
